sidebar category refactor
